Moved LastFMAPI from presentation to data.external.artist
